chore: rename src/test to src/test-fixtures and move organizations.test into src/server
