+++ src/auth-flow-contracts.test.ts
@@ -4,8 +4,8 @@ import { describe, expect, it, vi } from "vitest";
 import { AuthClient } from "./server/auth-client.js";
 import { StatelessSessionStore } from "./server/session/stateless-session-store.js";
 import { TransactionStore } from "./server/transaction-store.js";
-import { getDefaultRoutes } from "./test/defaults.js";
-import { generateSecret } from "./test/utils.js";
+import { getDefaultRoutes } from "./test-fixtures/defaults.js";
+import { generateSecret } from "./test-fixtures/utils.js";
 import { RESPONSE_TYPES } from "./types/index.js";
 
 const DOMAIN = "test.auth0.com";
