+++ src/route-defaults.test.ts
@@ -5,11 +5,11 @@ import { AuthClient } from "./server/auth-client.js";
 import { Auth0Client } from "./server/client.js";
 import { StatelessSessionStore } from "./server/session/stateless-session-store.js";
 import { TransactionStore } from "./server/transaction-store.js";
-import { getDefaultRoutes } from "./test/defaults.js";
-import { generateSecret } from "./test/utils.js";
+import { getDefaultRoutes } from "./test-fixtures/defaults.js";
+import { generateSecret } from "./test-fixtures/utils.js";
 
 vi.mock("next/headers.js", async () => {
-  const { createNextHeadersMock } = await import("./test/mocks.js");
+  const { createNextHeadersMock } = await import("./test-fixtures/mocks.js");
   return createNextHeadersMock();
 });
