+++ src/sdk-constants.test.ts
@@ -8,8 +8,8 @@ import { StatefulSessionStore } from "./server/session/stateful-session-store.js
 import { StatelessSessionStore } from "./server/session/stateless-session-store.js";
 import { TransactionStore } from "./server/transaction-store.js";
 import { DEFAULT_ID_TOKEN_CLAIMS } from "./server/user.js";
-import { getDefaultRoutes } from "./test/defaults.js";
-import { generateSecret } from "./test/utils.js";
+import { getDefaultRoutes } from "./test-fixtures/defaults.js";
+import { generateSecret } from "./test-fixtures/utils.js";
 import { RESPONSE_TYPES } from "./types/index.js";
 import {
   DEFAULT_DPOP_CLOCK_SKEW,
