+++ src/server/auth-client.test.ts
@@ -27,8 +27,8 @@ import {
   TokenRevocationError,
   TokenRevocationErrorCode
 } from "../errors/index.js";
-import { getDefaultRoutes } from "../test/defaults.js";
-import { generateSecret, stripTransactionValuePrefix } from "../test/utils.js";
+import { getDefaultRoutes } from "../test-fixtures/defaults.js";
+import { generateSecret, stripTransactionValuePrefix } from "../test-fixtures/utils.js";
 import {
   AccessTokenSet,
   RESPONSE_TYPES,
