+++ src/server/custom-token-exchange.test.ts
@@ -17,8 +17,8 @@ import {
   CustomTokenExchangeErrorCode,
   MfaRequiredError
 } from "../errors/index.js";
-import { getDefaultRoutes } from "../test/defaults.js";
-import { generateSecret } from "../test/utils.js";
+import { getDefaultRoutes } from "../test-fixtures/defaults.js";
+import { generateSecret } from "../test-fixtures/utils.js";
 import { SessionData } from "../types/index.js";
 import { TOKEN_TYPES } from "../types/token-vault.js";
 import { generateDpopKeyPair } from "../utils/dpopRetry.js";
