+++ src/server/fetcher.msw.test.ts
@@ -12,7 +12,7 @@ import {
   vi
 } from "vitest";
 
-import { createNextHeadersMock } from "../test/mocks.js";
+import { createNextHeadersMock } from "../test-fixtures/mocks.js";
 import { SessionData } from "../types/index.js";
 import { Auth0Client } from "./client.js";
 import { Fetcher } from "./fetcher.js";
