+++ src/server/fetcher.test.ts
@@ -1,8 +1,8 @@
 import * as oauth from "oauth4webapi";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 
-import { getDefaultRoutes } from "../test/defaults.js";
-import { generateSecret } from "../test/utils.js";
+import { getDefaultRoutes } from "../test-fixtures/defaults.js";
+import { generateSecret } from "../test-fixtures/utils.js";
 import { AuthClient } from "./auth-client.js";
 import { Fetcher } from "./fetcher.js";
 import { StatelessSessionStore } from "./session/stateless-session-store.js";
