+++ src/server/helpers/with-page-auth-required.test.ts
@@ -5,7 +5,7 @@ import { redirect } from "next/navigation.js";
 import ReactDOMServer from "react-dom/server";
 import { afterEach, describe, expect, it, vi } from "vitest";
 
-import { generateSecret } from "../../test/utils.js";
+import { generateSecret } from "../../test-fixtures/utils.js";
 import { Auth0Client } from "../client.js";
 import { RequestCookies } from "../cookies.js";
 import {
