+++ src/server/logout-strategy.flow.test.ts
@@ -11,8 +11,8 @@ import {
   it
 } from "vitest";
 
-import { getDefaultRoutes } from "../test/defaults.js";
-import { generateSecret } from "../test/utils.js";
+import { getDefaultRoutes } from "../test-fixtures/defaults.js";
+import { generateSecret } from "../test-fixtures/utils.js";
 import type { SessionData } from "../types/index.js";
 import { AuthClient } from "./auth-client.js";
 import { encrypt } from "./cookies.js";
