+++ src/server/mtls.flow.test.ts
@@ -22,8 +22,11 @@ import * as oauth from "oauth4webapi";
 import { beforeEach, describe, expect, it, vi } from "vitest";
 
 import { MtlsError, MtlsErrorCode } from "../errors/mtls-errors.js";
-import { getDefaultRoutes, setupMswLifecycle } from "../test/defaults.js";
-import { generateSecret } from "../test/utils.js";
+import {
+  getDefaultRoutes,
+  setupMswLifecycle
+} from "../test-fixtures/defaults.js";
+import { generateSecret } from "../test-fixtures/utils.js";
 import type { SessionData } from "../types/index.js";
 import { AuthClient } from "./auth-client.js";
 import { StatelessSessionStore } from "./session/stateless-session-store.js";
