+++ src/server/passkey/server-passkey-client.test.ts
@@ -15,7 +15,7 @@ import {
   createAuthorizationServerMetadata,
   getDefaultRoutes,
   setupMswLifecycle
-} from "../../test/defaults.js";
+} from "../../test-fixtures/defaults.js";
 import { AuthClientProvider } from "../auth-client-provider.js";
 import { AuthClient } from "../auth-client.js";
 import { encrypt } from "../cookies.js";
