+++ src/server/proxy-handler.test.ts
@@ -8,14 +8,14 @@ import {
   createAuthorizationServerMetadata,
   getDefaultRoutes,
   setupMswLifecycle
-} from "../test/defaults.js";
+} from "../test-fixtures/defaults.js";
 import {
   createDPoPNonceRetryHandler,
   createInitialSessionData,
   createSessionCookie,
   extractDPoPInfo
-} from "../test/proxy-handler-test-helpers.js";
-import { generateSecret } from "../test/utils.js";
+} from "../test-fixtures/proxy-handler-test-helpers.js";
+import { generateSecret } from "../test-fixtures/utils.js";
 import { generateDpopKeyPair } from "../utils/dpopRetry.js";
 import { AuthClient } from "./auth-client.js";
 import { StatelessSessionStore } from "./session/stateless-session-store.js";
