+++ src/server/redundant-txn-cookie-deletion.test.ts
@@ -15,8 +15,8 @@ import {
 } from "vitest";
 
 import { InvalidStateError, MissingStateError } from "../errors/index.js";
-import { getDefaultRoutes } from "../test/defaults.js";
-import { generateSecret } from "../test/utils.js";
+import { getDefaultRoutes } from "../test-fixtures/defaults.js";
+import { generateSecret } from "../test-fixtures/utils.js";
 import { SessionData } from "../types/index.js";
 import { AuthClient, AuthClientOptions } from "./auth-client.js";
 import {
