+++ src/server/session/stateless-session-store.test.ts
@@ -2,7 +2,7 @@ import { NextRequest } from "next/server.js";
 import * as jose from "jose";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 
-import { generateSecret } from "../../test/utils.js";
+import { generateSecret } from "../../test-fixtures/utils.js";
 import { CookieOptions, SessionData } from "../../types/index.js";
 import {
   decrypt,
