+++ src/server/transaction-store.test.ts
@@ -2,7 +2,7 @@ import * as jose from "jose";
 import * as oauth from "oauth4webapi";
 import { describe, expect, it } from "vitest";
 
-import { generateSecret, stripTransactionValuePrefix } from "../test/utils.js";
+import { generateSecret, stripTransactionValuePrefix } from "../test-fixtures/utils.js";
 import { RESPONSE_TYPES } from "../types/connected-accounts.js";
 import {
   decrypt,
