+++ src/test-fixtures/mfa-scenarios-shared.ts
